src/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
Celery==5.3.6
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1